01_docker_terraform/homework/pyproject.toml
[project]
name = "homework"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
